create `SchedulePlugin`
